+++ src/nvlib/model/ods/ods_writer.py
@@ -7,7 +7,7 @@ For further information see https://github.com/peter88213/novelibre
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 from string import Template
-from xml.sax.saxutils import escape
+from xml import sax
 
 from nvlib.model.odf.odf_file import OdfFile
 
@@ -348,7 +348,7 @@ class OdsWriter(OdfFile):
             entities['"'] = '&apos;'
         else:
             entities['"'] = '&quot;'
-        text = escape(text, entities=entities)
+        text = sax.saxutils.escape(text, entities=entities)
         return text.replace('\n', '</text:p>\n<text:p>')
 
     def _get_extra_styles(self, elements):